+++ HWDescription/FEDescription.h
@@ -1,7 +1,7 @@
 #ifndef FEDescription_h__
 #define FEDescription_h__
 
-#include "Rtypes.h"
+#include <cstdint>
 
 // Base Class to describe all parameters common to all FE Components in the DAQ chain
 // mainly connections & staus for the moment
@@ -15,21 +15,21 @@ namespace Ph2_HwDescription{
 		// MEMBERS
 
 		// Crate (uTCA) that the FE is connected to
-		uint32_t fShelveId;
+		uint8_t fShelveId;
 		// BIO Board Id that the FE is connected to
-		uint32_t fBeId;
+		uint8_t fBeId;
 		// Id of the FMC Slot on the BIO Board, all FEs need to know so the right FW registers can be written
-		uint32_t fFMCId;
+		uint8_t fFMCId;
 		// Id of the FE (module/hybrid, etc...)
-		uint32_t fFeId;
+		uint8_t fFeId;
 		// status (true=active, false=disabled)
 		bool fStatus;
 
 		// METHODS
 
 		// 3 C'tors with different parameter sets
-		FEDescription( uint32_t pShelveId, uint32_t pBeId, uint32_t pFMCId, uint32_t pFeId, bool pStatus=true );
-		FEDescription( uint32_t pBeId, uint32_t pFMCId, uint32_t pFeId );
+		FEDescription( uint8_t pShelveId, uint8_t pBeId, uint8_t pFMCId, uint8_t pFeId, bool pStatus=true );
+		FEDescription( uint8_t pBeId, uint8_t pFMCId, uint8_t pFeId );
 		FEDescription( );
 
 		//Copy C'tors
@@ -39,26 +39,26 @@ namespace Ph2_HwDescription{
 		virtual ~FEDescription( );
 
 		// Getter methods
-		 virtual uint32_t getShelveId()
+		 virtual uint8_t getShelveId()
 		{return fShelveId;};
 
-		 virtual uint32_t getBeId()
+		 virtual uint8_t getBeId()
 		{ return fBeId ;};
 
-		 virtual uint32_t getFMCId()
+		 virtual uint8_t getFMCId()
 		{ return fFMCId;};
 
-		 virtual uint32_t getFeId()
+		 virtual uint8_t getFeId()
 		{ return fFeId ;};
 
 		 virtual bool getStatus()
 		{ return fStatus ;};
 
 		// Setter methods
-		 virtual uint32_t setShelveId( uint32_t pShelveId );
-		 virtual uint32_t setBeId( uint32_t pBeId );
-		 virtual uint32_t setFMCId(uint32_t pFMCId);
-		 virtual uint32_t setFeId( uint32_t pFeId );
+		 virtual uint8_t setShelveId( uint8_t pShelveId );
+		 virtual uint8_t setBeId( uint8_t pBeId );
+		 virtual uint8_t setFMCId(uint8_t pFMCId);
+		 virtual uint8_t setFeId( uint8_t pFeId );
 		 virtual bool setStatus (bool pStatus );
 	};
